Replace string-based typecheck with class-object based typecheck.
This appears to avoid the -G bugs for Rakudo, which appear to occur
due to something going wrong when do_dispatch (written in C) calls
!bindability_checker (PIR) which calls typeof_s_p opcode which
calls the 'name' vtable function which itself happens to be
overridden in PIR via a :vtable flagged sub.  Or something like that.
Anyway, we once again see that using string names as class identifiers
leads to trouble.  (Bad coder!  No biscuit!)