remove two junction related tests from spectest.data
One was deleted, the other one now assumes we can call $j!eigenstates, which
is not so easy, and needs some magic and scary preamble, which we don't know
yet. So from a Rakudo perspective that test is dead, but since it's based on a
spec change that's OK for now.