Rework Range and RangeIterator to properly handle Ranges which exclude their ends and non-numeric Ranges.
In the process, move Range!max_test to Range.max_test so RangeIterator can use it.  There's probably a better way to do this, but this is the best I know how to do, and it works.  (I think.  I'm loath to spend much time defining RangeIterator tests until the Iterator standard is a bit more stable.)