Change master spectest to use t/spec revision 29188.
Because Rakudo development has shifted from master to ng for the moment, while some significant changes are being made to the spec, we are setting master to point to a fixed older release of the spectests.  This means we can make changes to the spectests without creating "regressions" in Rakudo master.