[Set.pm] new class in core
It currently does union, intersection and set difference. It contains methods
for checking subset and superset relationships. There are two sets of ops for
the above: one unicode set and one 'Texas' parenthesized set. This is all a
bit above and beyond the current spec, but quite close to what I envision for
core.