[release] bump parrot revision to 2.10.0 release
Note that parrot 2.10.0 has a known bug that prevents it from being installed
if git is not installed. That's unfortunate, and a 2.10.1 release exist that
fixes it, but it's nowhere to be found in the git repo.

gerd++ promised to put the stuff into the parrot repo soon, when that's done
PARROT_REVISION should be bumped to the new tag