workaround for RT #86906 - .panic fails to panic in some grammar rules
The fix should really be to make the context information available, but since
the control flow is a bit weird in that parts of the grammar, I don't know how
to do this. Instead it's just s/self\.panic/pir::die/, which improves the
error message, but makes line numbers unavailable.