Remove invalid Pod
S26 states:
After the typename, the rest of the C<=begin> marker line is treated as
configuration information for the block. This information is used in
different ways by different types of blocks, but is always specified
using Perl6-ish option pairs.

We don't parse the block configuration information yet, plus they're not
needed anyway there.