implement infix<+>(Any, Any)
This is the same way as in master, through a .Bridge method
which converts to a common numeric type if no direct multi
is found.

I don't quite like the current factoring, because something
like "3" + 4.4e0 goes through (Any, Any), then (Real, Real) and finally
through (Num, Num) - three dispatches for a single, coercive multi.

Any suggestions on how to improve it (and still keep it extensible
when the user adds enw Real types) are very welcome.