Add pir::const:: for passing pasm constants to PIR.
Note that these only work within pir::op(...) statements and other
items -- you can't assign a pir::const:: value directly to a Perl 6
variable.  Binding with C<:=> might be possible, but in that case
you're getting a Parrot Integer PMC and not a Perl 6 Int.  If/when
we need the capability for pasm constants to appear directly in
our code, we'll introduce a rakudo::const or similar namespace for it.