Dynamic vars -- stub in dynamic variable lookups.
This only handles the lookup portion of dynvars; as yet dynamic
variable declarations (my $*abc) don't work and afaik there's not
a way to create package-scoped variables (e.g. in PROCESS).
But the stub code is here anyway.   (Use pir::say($*AUTOPRINTPOS)
in interactive mode to see that dynvars are actually searching the
dynamic context. :-)