More Str.Numeric fixes and improvements.
Although converting a string with no number is supposed to fail,
doing so causes all sorts of issues throughout the spectests with
no clear-cut solutions.  Therefore I'm fixing the conversion to
treat non-numbers as integer zero for now, until we can regroup
and figure out how we want to handle the failure cases.

Note that any non-whitespace characters following a valid number
or partial number still results in a failure, though.