Bring back Perl 6 level stack traces
There are still a lot of quirks to be worked out, for example some routines
magically not appearing in the stack trace, and other showing up
that I would prefer to keep out (like &die), but I believe it is a net
improvement. The old PIR backtraces can be obtain with the --ll-exception
command line argument.