Cleanup exception handling, so blocks with handlers can also be inlined.
Instead of perl6_rethrow_skipnextctx op, we now have the much nicer
perl6_based_rethrow, which rethrows an exception in the context of
another exception. To make blocks with both CATCH and CONTROL handlers
work, we need another op to skip the "outer" handler if an exception
is caught in the exception handler. I.e. a die() in a CONTROL handler
must not be caught with a "parallel" CATCH handler.