improved backtrace printer
When encountering a a non-Routine, it searches for an outer Routine that
also appears in the call chain, and takes their name, but the line number
of the block. The idea comes from jnthn++

Testing welcome, not sure yet how well it interacts with lazy lists. At
least for normal for-loops it cuts all those eager/reify methods.