The last patch almost worked but not quite.  The old copy of the announce directory copy wrongly expanded as below:
/bin/cp     README CREDITS LICENSE docs/*/announce    /home/ron/rakudo/install/share/doc/rakudo/announce

and got a stat error on docs/*/announce and failed.  One might also remove the nearby announce lines but if the announcements are wanted my patch will put them there without crashing.  In the rakudo build this crash does relatively little harm since it is near the end of the install.  In rakudo star the crash stops the process before module builds.