Revise Version.Str / .gist / .perl .
After thinking about it a bit more, I'm of the opinion that
$version.Str should produce a string that doesn't include
a leading 'v'.  The .gist form does have a leading 'v', however,
so C< say $version > still does what one expects.

When putting a version in the middle of a string we'll often want
to exclude any leading 'v' (e.g., when generating filenames),
it's much easier to add a 'v' when it's needed than it is to
eliminate one we don't want.  Also, the form without the
leading 'v' is the only one that can (so far) reliably have
much chance of round-tripping back into Version.new(...),
since alphabetics in strings arguments to Version.new are meaningful.

At the moment the spec is somewhat silent on this topic, so
I'm hoping this will make more sense.  If not, a spec
clarification or some official tests will let us know how to
improve it.  :-)