Proper handling of sorry and worry.
Makes worry just report the potential difficulties (so redeclaration
of a variable is just a warning now, as STD has it), and sorry keeps
going and just collects together the errors up to a limit, then gives
them at the end or when there's a panic. Slightly different factoring
to STD since we have typed exceptions, but produces the same kind of
output.