Start recording mystery symbols.
Factored in a similar way to %*MYSTERY in STD. However, it keyed just
on name. This isn't quite enough, as it fails to distinguish the same
names in different scopes, meaning mysteries can be resolved wrongly.
To avoid this, we'll try keying on name and scope ID in Rakudo, so we
collect together obvious dupes, but hopefully can get righter results.

Information not used for anything just yet.