Don't re-link already compiled blocks.
This loses lexical chain information they are carrying. A test for
traits/wrapping/multi interaction (yes, we have one :-)) seems to have
been somehow magically working before, for mysterious reasons. This
change makes it work for a better understood reason, and with the NQP
based multi-dispatcher. Fixes the remaining regression in this branch,
and introduces no more spectest regressions.