Yet more exceptiony stuff.
Heck knows if this'll work. But it probably will; the exception stuff
in the NQP on JVM runtime runs handlers in the dynamic scope of the
throw by default and already takes care not to let active handlers
catch exceptions thrown within them, which it seems is what the dynops
we avoid using here do. Caveat: this isn't a solution for LEAVE.