Make p6capturelex not create a busted chain.
It didn't care what it shoved into .outer, meaning that we could put
entirely wrong things in there. This tries to do a bit better, though
really the longer-term solution is to fix up the code-gen, hopefully
by leaving the QAST compiler to work it out. In the meantime, this
seems to fix various lexical issues.