Start replacing Buf.
Buf is now a role, as per spec. We also provide buf8, buf16, buf32 and
buf64 symbols in the setting, which are instantiations of it. However,
VMArray isn't quite smart enough to pack those as it should yet. Much
previous Buf functionality is still missing; it will be added back
before this branch is merged.