Reimplement :with for List.(uniq|squish) (keep :as available)
Yes, it was removed from the spec, but the alternative for doing [eqv] based
uniqueness using *.perl is 1. still not possible now (because .perl doesn't
always give an accurate representation), and 2. very resource intensive,
because a whole structure needs to be stringified first before being able to
find out that they're different.  Whereas [eqv] is very quickly able to find
out whether two structures are different.

So, do I believe it belongs in the spec?  Yes!  :-)