Support rakudo-jvm builds on smaller systems
Tune JVM memory options to allow successful compile on a
32-bit/2GB RAM system, while still remaining <2% slower to
compile the setting on a larger 64-bit system than the current
settings.

Max heap size was tested every 100m; -Xmx1100m was minimal for
compiles on 32-bit, and -Xmx1300m was minimal on 64-bit, but
noticeably slower than current settings.  -Xmx1600m seemed to
be the sweet spot for minimal slowdown on 64-bit while still
being buildable on a 32-bit/2GB RAM system.

Also explicitly set stack size, as the automatic tuning seemed
to work fine on the larger 64-bit system, but set a too-small
stack on the smaller system, causing the build to fail.
-Xss1m seems to work well on both.