Simplify DEPRECATED / is DEPRECATED for now
Before, I was thinking of having sub DEPRECATED figure out what is was warning
for, with using callframe().  But this proved to be very fragile.  So instead,
sub DEPRECATED now just gets the old and new name, and the "is DEPRECATED"
trait adds a closure in the form of once block in an ENTER phaser for the
routine.  For this to work inside routines without explicit ENTER blocks,
the time that "finish_code_object" in Actions.nqp is called, must be changed.
But I'll leave that for tomorrow.