Use a self-annihilating ENTER phaser for "is DEPRECATED" trait
This is again a proof of concept.  Because of &?BLOCK not existing, it is
impossible to create a Block.remove_phaser($type,&?BLOCK) capability.  Instead,
this commit creates a "pop_phaser" method that will simply pop the last phaser
of the given type from the list of phasers for that type.  Since the ENTER
phaser of the "is DEPRECATED" trait is added *after* any other ENTER phasers in
that block, popping a phaser means removing the one that was added for
"is DEPRECATED".  Maybe we need a better way for a phaser to remove itself
from future execution.