Stub in p6bindsig.
Due to the many calls the binder makes back into the interpreter, we
will just write it in NQP code on Moar. This will be a little slower
at first but (a) we expect Moar to get faster, and (b) we expect to
start "lowering" simple signatures down to QAST::Var usage, which is
also going to be useful for making better use of invokedynamic on the
JVM backend.