HashIter::reify now does eager snapshot
There is some instability in the reentrancy of hashiter, so we force
reify to run to completion in lieu of a bug fix.  But hashes are
finite anyway, and not many people are going to want to do short-circuit
lazy evaluation of a list of keys or values, so not a big deal if we
don't support laziness on hashes for now.  This fix enables maps to
reify their input lists frugally rather than greedily, since greedy map
was hiding the reentrancy problem that this patch hides differently.