MapIter::reify is now frugal with input list
Just because a map is in eager context doesn't mean that its input
list should be evaluated eagerly, since loops can terminate early.
We also want to be able to pull lazily from an infinite input list.
Regardless of the demand, we only need to make sure each call into
the closure has just enough values to bind.  Also, there's no point
in preallocating the output rpa when it will be unused in sink context.