Range::reify decides range batching now
So that Map::Iter doesn't have to fake batching decision with
insufficient knowledge.  For now we batch up to 1024 values; a more
sophisticated decision based on max - min is possible, in addition
to any other contextual clues that might indicate one-shot use in
cool code vs repeated use of the range in hot code.