list infix reductions no longer flatten
Since we fixed X and Z metaops to not violate Scalar containers,
we have to be able to pass a list of parcels to [Xop] and [Zop]
and have each parcel treated as a separate list.  The triangle
forms also work (slowly, since they have to recalculate each stage
from the start), but there is an unbox warning I don't understand yet.