Fix sub calls from pre-compiled roles in Moar.
In some cases, the fixup to the role outer would not be done. On the
Parrot and JVM backends, things "magically" work out due to keeping
track of a prior invocation. Moar's stricter closure model (which the
JVM backend should also move to) lacks the "magic", and so highlighted
the reliance on it. This fixes things for on MoarVM, but leaves alone
JVM and Parrot's current approach given we're right up against the
monthly release; after the release, they can all switch to the same,
newer, model.