more robust error reporting
If a thing is not a string it is not automatically an Exception,
so check nqp::can($possible_ex, throw)

Also be more liberal in what the subs in P6EX accept. In parrot land,
we mostly passed around type names, but in moar we use actual types,
so the signatures must accept Mu

Junction authothreading still seems broken; I hope timotimo++ fixes it.