Assorted updates now that onlystars are thunks.
This fixes many long-standing inconsistencies when we skip over any
onlystar protos. The multispec optimization forced the issue, but it
has existed for a couple of years, since we started inlining or doing
multi-sub compile time call optimization. Note, updates will be needed
for the JVM and Parrot backends to cope with these changes.