Use .perl instead of .WHICH on Capture as key
Suggested by masak++.  This will however cause cache misses for named parameters
as the keys in Hash.perl are not sorted.  Perhaps they should be?  Even though
it is not necessary for roundtripping per se, SortKeys *is* one of the features
I always used to switch on on Perl 5's Data::Dumpern when debugging.