Fix pod directives formatting codes in delimited code blocks.
This makes handling of pod directives in code blocks consistent whether
or not you've wrapped it in a multi-line (and :allow'd) formatting code.

In delimited code blocks, the only pod directive that can't be stuffed
in formatting codes is the same one that can't be in a delimited code
block in the first place: =end code. For other code blocks and that pod
directive in particular, you'll need to keep = from being the first
non-whitespace character, same as before.