Very first pass at signal handlers.
Probably some we could support are missing (though it's an LHF for
somebody to add those), and the API is certainly up for discussion.
But now there is, at least, a way. Closing the tap doesn't in any
way remove the handler yet; that'll come later once the more general
cancellation mechanism arrives.