Properly handle closing a tap in many built-ins.
This was a leftover NYI from when supplies were first added. Now you
can close a tap right at the end of a chain and it cascades all the
way up to the top. For example, if there's a Supply.interval at the
end of the chain, it will stop the underlying timer from producing
values.