Implement start method on supplies.
Takes a closure and, for each supplied value, schedules the closure to
run on another thread. It then more's a Supply (resulting in us having
a supply of supplies) that will either have a single value more'd and
then be done if the async work completes successfully, or quit if the
work fails. Useful for kicking off work on the thread pool if you do
not want to block up the thread pushing values at you (maybe 'cus you
are reacting to UI events, but have some long-running work to kick
off).