remove bogus literal complex rule
We should not have a special literal rule for 42i, for several reasons:
First, it's incorrect in the face of language extensibility, since
someone could add a different affix that overrides i in either
precedence or semantics.  Second, the rule was incorrect anyway, since
it didn't check for a word boundary after the i.  Third, we can and
should rely on constant folding these days to reduce 42i to a literal
value when it's appropriate (and refrain from doing so when it's not).
Fourth, a literal complex ought to be written with angles anyway,
if you don't want the constant folder to get involved.  That's why
that construct is there.  Fifth, this fixes the double warning on
(non)-octals beginning with 0, since the complex rule was calling the
integer rule and then backtracking in the absence of an i.