implement labeled loops and throwing of labels as payload
Labels are lexcially installed as instances of type Label, and are glued to the
exception handlers of e.g. 'while', 'until' and 'for' loops. When an exception
with a label as payload is thrown, the handler that holds the identical label
will be in charge. Introspecting the labels of a block is NYI, since we do not
have &?BLOCK to yet, see S06:3047.