fix precomp bug about colonpairs in 'use' statement
When we pass a colonpair to a use statement it can land in the compiled
compilation unit before the CU's serialization context is created. Which
means that we cannot have WVal there, since these need the SC of the CU
to be deserialized. This patch cares about strings, i.e. in :from<NQP>.
We need a smarter solution in future to be able to use regexes, versions
and other objects in colonpairs there.