Remove all lines that match ^# \w
These are all comments that we don't need in the concatenated setting (I don't
think), at least not right now when we don't have line numbering going back to
the constituent modules.  Hopefully, this will mean that there's less of a
setting building penalty for adding some internal documentation to the setting.