Make block inlining a level 2 optimization.
It's quite worthwhile; we've had it as a level 3 optimization for a
long while now, applying it to the setting. perl6-bench also used it.
For a while, lack of OSR in MoarVM meant it could make things worse in
the odd case; now that's dealt with, and the issues that caused it to
blow up a few spectests have been isolated, we'll try enabling it as
a default optimization. Plenty of time up to release for testing.