p6bindsig should not prevent immediate inlining
While on MoarVM we can lower signatures and thus optimize this out,
on other backends we don't, but should not lose the immediate inline
optimization on those. FROGGS++ also tracked down a behavior change
that resulted from disabling this optimization on r-p; it should not
make a difference, but this does indeed fix the r-p breakage.