Tweak some protos for better Junction failover.
We can make the proto itself be (Any, Any), meaning it fails to bind.
However, this also means that the multi-dispatcher never gets chance
to sneak an entry into the cache for the Junction case, so we have to
go through the slow-path binder to fail over every time, rather than
cheaply identifying we should hit the Junction dispatcher.