Almost complete refactor CompUnit and CURL::File
A CompUnit object now has always the source file as its "path", even if the
source file cannot be found.

CURL::File has been utterly simplified, as it turned to be impossible to
smartmatch finding files.  So now it only looks for the existence of the
file, and returns that.  Thanks to Scalar[0] indexing, this is compatible
with the API.