just do a straight eager rpa write, 10% faster
No point in involving gather/take at all when we know this is eager.
Could probably speed this up further if we had low-level memory copying
of rpa pointer data, so we could batch the memory initialization.

Actually, this runs 10% faster on Moar, but a little slower on JVM than
previous.  We'll optimize for the better optimizer at this point, which
is Moar's.  Presumably the JVM's optimizer will catch up sooner or later.