move LOLLY hack from parser down into harness
This causes our new version to default to LOLLY without having to set
LOLLY=1, but the test harness now forces LOLLY=1 for fudging, so we're
testing against the new semantics.  At the same time, the latest tests
can still be used by old versions for now if that is desired, since
the harness is version specific and the old harness doesn't set LOLLY.
In a month or so we can delete all mention of LOLLY from the harness
and from roast.  But no harm if we forget. :)