Say what you expected before you say what you got!
Expectations happen before the operation, what you got happens as a
result of the operation.  Saying them in the wrong order has bugged me
for years, but for some mystical reason I'm usually distracted by some
other bug when I see a test failure, and keep forgetting to fix this.