Hack a Parcel.WHICH solution
Apparently, a Parcel created by the .Parcel method, consists of a ListIter
as its only element.  Upon inspection, this is flattened, causing a problem
with the .WHICH representation, as the immutable Parcel internally mutes.
This fix will just expose the ListIter if the Parcel was created with
.Parcel, and thereby create a WHICH that indicates the inherit mutability
of a Parcel that is created that way.  Suggestions for a better solution
are welcome.