Bulid a * and ** during compile, and use them.
This means that we eliminate the method calls to .new in both cases,
as well as a bunch of object allocations. This should - at least in
the pre-compiled module case where the created object is in gen2, or
when the compilation ran long enough to promote it - JIT down to an
instruction or two on Moar. Even if it's the less direct lookup, it
will still be enormously faster than a method call.