Improve take[-rw], proceed, succeed.
By making them multis, the common cases can avoid a Parcel allocation
along with the underlying list construction, as well as a call. This
takes around 5% off a "read a million lines" benchmark, and probably
is more significant in things that hot-path on gather. The succeed
improvement should help all given/when usages.