Wire up DESTROY (sub)methods to finalize mechanism
Since it's only supported on MoarVM so far, that's the only backend
you'll see it working on at the moment. As usual with these things,
there's no promises about when the finalier will be called, and if
the program ends before the object is GC'd it won't be called at
all.