Don't fatten args to Bag.new/Mix.new.
They already slurp, so there's no reason to do this. In fact, there's
a very good reason not to: flattening the args may produce a huge
number of them, which is painful performance wise and hits limits
on JVM and on MoarVM too. This appears to fix the issue reported in
RT#122810.