First part of cruft removal from IO::Handle
Because before, .IO would give you an unopened IO::Handle.  So any time you
would call a method on an IO::Handle that would require an open file, it would
have to check whether the file was open already or not, and whether it was a
directory or not.  Because it is now a lot harder to create an unopened
IO::Handle, we can drop the tests now.

This should particularly make loops with .get and .getc a *lot* faster.