Have $a.^foo($b) desugar to $a.HOW.foo($a, $b).
Before it desugared to .dispatch:<.^>, which would in turn pun roles -
not to mention be a decent bit more costly and not work out on things
that hadn't had their inheritance set up yet. Most patches before this
were clearing up things that accidentally relied on the previous way
things worked with punning roles.