Fix .perl for typed arrays
This is really fixing a symptom rather than the cause, because the real problem
is that the iterator on the typed array is returning the wrong thing.  Since
this gets me into gimme and reify land, I opted for this solution, so that at
least .perl roundtrips for typed arrays.