speed up /$string/, fix /$array/ to stringify
INTERPOLATE was setting up a very slow loop to traverse a single item.
It was also depending on $order sometimes being a Perl 6 value and
sometimes being an NQP value.  It was also incorrectly introspecting
positionals that are supposed to be hidden inside of scalars.

This method is still doing far too much introspection and should be
split up into more specialized routines.