Introduce filetest and abspath primitives
At one point we probably want to have these live in a role.  But at the
moment it is unclear to me how that role would work in the different classes.
So for now, I create these as primitive subs.  And maybe they will stay this
way, because they are better optimizing / inlining targets this way.