Fix IO::Spec::Cygwin.rel2abs on Windows.
The tests (I think correctly) don't expect to end up with a volume,
but because we delegate to Windows rel2abs on Cygwin then tweak the
slashes, we end up with one. Not sure this is the best fix, though it
is better than the other obvious alternative (let the drive letter be
put on by IO::Spec::Win32, then strip it off in IO::Spec::Cygwin).