Fix bug in PARSE-INCLUDE-SPEC
For some reason, the $class was pass by reference instead of value.  Which
caused all of the entries to have the final class found, not the actual class
active at the moment of pushing.  Not sure if this is a bug in general, a bug
in the setting, or a bug in my head.