Allow specification of @*INC in CompUnit.precomp
By default, CompUnit.precomp will now *only* use whatever is in @*INC (by
default) to set up the @*INC in the external process doing the precompilation.

This actually breaks a few spectests (and actually passes a todo test), but
does not break panda (mostly because it doesn't use .precomp yet).  Hopefully
this will make it easier to trace some hard to find precomp related bugs and
other oddness.  If not, then please revert this one.