Recommend attributes in addition to similar lexicals
We currently have a mechanism for suggesting similarly named lexical
variables if someone mistypes a variable name:

    my $name = 'Rob';
    say "Hello, $nam!";

    # output: Variable '$nam' is not declared. Did you mean '$name'?

I personally often make the mistake of typing '$name' when I mean to
refer to the member of an object '$!name'.  This commit adds $!name
as a suggestion if you're in a context that has member variables
and has a member variable $name.