Re-organize parametric role group handling.
We want to create a single point of interning in order to make use of
the 6model parametric extensions. Dis-entangle resolving a role group
to a particular role from the specialization of the chosen role, so we
will be able to apply the type interning inside of the parametric role
type's meta-object. (This may, as a side-effect, enable us to use the
group type as a point of interning curries.)