Keep compiler callbacks in place during optimizer.
The optimizer may trigger compilation of things that it calls - in
the setting because it defines many built-ins, but also in normal
code that uses, for example, 'is pure'. This prevents us from
removing fixup infrastructure too early. A pre-requisite for fixing
problems with multis that get run during compilation when a subclass
has derived the proto.