Track and update clones of uncompiled coderefs.
This happens if a proto method in a base class is cloned as part of
deriving a proto for a subclass, and afterwards the proto in the
base class was invoked before the compilation process was over. The
lack of update led to a problem locating the appropriate static code
object during serialization.