Make adverbed slices up to 10% faster still
Some backstory: slicing was done by a large piece of code that handled both
array slices and hash slices.  At various places, a flag was checked to do
either for instance an "at_pos" or an "at_key".  By creating a P6 script
in tools/build (makeSLICE.pl6) that templates that code to generate separate
cases for list and hash slices, we effectively removed a lot of flag passing
and checking.  The generated code should be better optimizable / jittable.
At the expense of a larger core setting, of course.  Which is the reason I've
done this as a single commit, so that we can revert easily if needed.