add optional digits arg to .base
Note that we add the argument even to Int.base because we want to be
robust in the face of type narrowing.  it's just that the fractional
digits on an integer, if requested, are always going to be 0s.

Also, Real.base was not correctly rounding in the last place.