A lot of work related to directories
In the end, it turned out not to be such a good idea to make IO::Dir have its
abspath with a trailing /.  So, removed that: IO::Dir now only stringifies with
a trailing /.  This had quite some repercussions, hope I caught them all.

Also implement IO.resolve (on Moar), now that we nqp::readlink.