add typed Pointer type as replacement for OpaquePointer
A Pointer can still be used as is, and will Pointer[void] (void*).
When a Pointer points to something more useful(i.e. Pointer[int32]),
one can call method 'deref' on the pointer to reveal the the thing
it is pointing to. Besides that a Pointer can still be instantiatet
using a memory address and numifies back to that mem address.
We export the OpaquePointer type as an alias to Pointer, to allow
adoption before we OpaquePointer gets deprecated.