keywords now require whitespace (unless shadowed)
Now we parse keywords more like STD, requiring whitespace.  When we parse
what might be a keyword, if there's whitespace after, it's always taken
as a keyword.  If not, if the term has been defined elsewhere, we take
that other definition.  Otherwise we carp about missing whitespace after
the keyword.
    my \if = 42; say (if) if if; # now says 42